PathArgChecker instantiated after the path is obtained (#4)
Class MissingPathArgWarner warns the programmer if a path argument is missing. If the path is provided, this new class can instantiate a PathChecker or a PathArgChecker. Therefore, it eliminates the need to instantiate PathArgChecker before the path is obtained. Class ArgPathChecker was renamed PathArgChecker because the documentation tells about path arguments, not argument paths.